Adicionando funções específicas

+++ backend/app/controllers/AssinaturaController.py
@@ -1,21 +1,26 @@
 import uuid
+from flask_jwt_extended import get_jwt_identity
 from backend.app.models.Assinatura import Assinatura
 from datetime import datetime
 from backend.app.db.config import db
 
-def listar_assinaturas(id_usuario):
+def listar_assinaturas():
+    id_usuario = get_jwt_identity()
     assinaturas = Assinatura.query.filter_by(id_usuario=id_usuario).all()
     return [a.to_dict() for a in assinaturas], 200
 
 def buscar_assinatura_por_id(id_assinatura):
+    id_usuario = get_jwt_identity()
     assinatura = Assinatura.query.get(id_assinatura)
     if not assinatura:
         return {'mensagem': 'Assinatura não encontrada.'}, 404
+    if assinatura.id_usuario != id_usuario:
+        return {'mensagem': 'Acesso negado: esta assinatura não pertence a você.'}, 403
     return assinatura.to_dict(), 200
 
-def criar_assinatura(data, id_usuario):
+def criar_assinatura(data):
+    id_usuario = get_jwt_identity()
     try:
-        # Converte as datas de string para objetos datetime.date
         data_inicio = datetime.strptime(data.get('data_inicio'), '%Y-%m-%d').date()
         data_fim = datetime.strptime(data.get('data_fim'), '%Y-%m-%d').date()
 
@@ -36,13 +41,19 @@ def criar_assinatura(data, id_usuario):
         return {'mensagem': f'Erro ao criar assinatura: {str(e)}'}, 500
 
 def atualizar_assinatura(id_assinatura, data):
+    id_usuario = get_jwt_identity()
     assinatura = Assinatura.query.get(id_assinatura)
     if not assinatura:
         return {'mensagem': 'Assinatura não encontrada.'}, 404
+    if assinatura.id_usuario != id_usuario:
+        return {'mensagem': 'Acesso negado: esta assinatura não pertence a você.'}, 403
+
+    if 'data_inicio' in data:
+        assinatura.data_inicio = datetime.strptime(data['data_inicio'], '%Y-%m-%d').date()
+    if 'data_fim' in data:
+        assinatura.data_fim = datetime.strptime(data['data_fim'], '%Y-%m-%d').date()
 
     assinatura.tipo_assinatura = data.get('tipo_assinatura', assinatura.tipo_assinatura)
-    assinatura.data_inicio = data.get('data_inicio', assinatura.data_inicio)
-    assinatura.data_fim = data.get('data_fim', assinatura.data_fim)
     assinatura.status = data.get('status', assinatura.status)
     assinatura.preco_assinatura = data.get('preco_assinatura', assinatura.preco_assinatura)
 
@@ -50,9 +61,12 @@ def atualizar_assinatura(id_assinatura, data):
     return {'mensagem': 'Assinatura atualizada com sucesso!', 'assinatura': assinatura.to_dict()}, 200
 
 def deletar_assinatura(id_assinatura):
+    id_usuario = get_jwt_identity()
     assinatura = Assinatura.query.get(id_assinatura)
     if not assinatura:
         return {'mensagem': 'Assinatura não encontrada.'}, 404
+    if assinatura.id_usuario != id_usuario:
+        return {'mensagem': 'Acesso negado: esta assinatura não pertence a você.'}, 403
 
     db.session.delete(assinatura)
     db.session.commit()
